use canned CSV files instead of trying to run the profiler
