+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_cbCD7a/rapids_4_spark_profile/app-20231122005806-0064/io_metrics.csv
@@ -1,0 +1,104 @@
+appIndex,appID,sqlID,input_bytesRead_sum,input_recordsRead_sum,output_bytesWritten_sum,output_recordsWritten_sum,diskBytesSpilled_sum,memoryBytesSpilled_sum,sr_totalBytesRead_sum,sw_bytesWritten_sum
+1,"app-20231122005806-0064",24,288435611,28828293,0,0,0,0,4012,4012
+1,"app-20231122005806-0064",25,159918237,7514061,0,0,0,0,2569432,2569432
+1,"app-20231122005806-0064",26,924109967,24807134,0,0,0,0,334837750,219497829
+1,"app-20231122005806-0064",27,144601149,102000,0,0,0,0,0,0
+1,"app-20231122005806-0064",28,23095855,4222010,0,0,0,0,71162592,70106037
+1,"app-20231122005806-0064",29,23095855,4222010,0,0,0,0,69140811,69043970
+1,"app-20231122005806-0064",30,496406884,6261340,0,0,0,0,151055752,151055752
+1,"app-20231122005806-0064",31,29944401,938110,0,0,0,0,4022861,4022861
+1,"app-20231122005806-0064",32,20289368,1751893,0,0,0,0,541461,541461
+1,"app-20231122005806-0064",33,74342598,2161228,0,0,0,0,39779873,39779873
+1,"app-20231122005806-0064",34,692385530,14810644,0,0,0,0,523745805,523745805
+1,"app-20231122005806-0064",35,56923258,1556431,0,0,0,0,160800,160800
+1,"app-20231122005806-0064",36,76011313,1631108,0,0,0,0,25578638,25578638
+1,"app-20231122005806-0064",37,18837333,2717555,0,0,0,0,21203,12012
+1,"app-20231122005806-0064",38,26199183,4712010,0,0,0,0,5096333,5096333
+1,"app-20231122005806-0064",39,67913119,5539245,0,0,0,0,30701,30701
+1,"app-20231122005806-0064",40,160305405,7547528,0,0,0,0,5080043,5080043
+1,"app-20231122005806-0064",41,564079685,15384391,0,0,0,0,221612321,221612321
+1,"app-20231122005806-0064",42,18061414,1100405,0,0,0,0,3494,3494
+1,"app-20231122005806-0064",43,5717344,376768,0,0,0,0,1841920,1841920
+1,"app-20231122005806-0064",44,988950033,32585500,0,0,0,0,940879932,888481751
+1,"app-20231122005806-0064",45,115358381,5628926,0,0,0,0,194652,194652
+1,"app-20231122005806-0064",46,17964560,1021793,0,0,0,0,851781,851781
+1,"app-20231122005806-0064",47,90891333,5553575,0,0,0,0,3963418,3963418
+1,"app-20231122005806-0064",48,212905544,7217638,0,0,0,0,2319863,2319863
+1,"app-20231122005806-0064",49,832682945,30512295,0,0,0,0,422443064,422443064
+1,"app-20231122005806-0064",50,44315300,4624328,0,0,0,0,9958076,9958076
+1,"app-20231122005806-0064",51,91772331,5634634,0,0,0,0,282039,282039
+1,"app-20231122005806-0064",52,36458461,4060171,0,0,0,0,9671209,9671209
+1,"app-20231122005806-0064",53,26629503,1570270,0,0,0,0,1942700,1942700
+1,"app-20231122005806-0064",54,337035976,27618067,0,0,0,0,5683078,3039100
+1,"app-20231122005806-0064",55,209235761,18603261,0,0,0,0,73077371,73077371
+1,"app-20231122005806-0064",56,11918305,392382,0,0,0,0,14298679,12518446
+1,"app-20231122005806-0064",57,192095309,6263946,0,0,0,0,29757039,29757039
+1,"app-20231122005806-0064",58,115613150,11045380,0,0,0,0,14862,14862
+1,"app-20231122005806-0064",59,109277126,5635115,0,0,0,0,750123788,750123788
+1,"app-20231122005806-0064",60,3623155560,172805846,0,0,0,0,22335865,22335865
+1,"app-20231122005806-0064",61,69117633,1592350,0,0,0,0,36258615,36258615
+1,"app-20231122005806-0064",62,95551722,8314423,0,0,0,0,96537511,96537511
+1,"app-20231122005806-0064",63,202759204,4394020,0,0,0,0,114066,114066
+1,"app-20231122005806-0064",64,398831294,14462532,0,0,0,0,7198,7198
+1,"app-20231122005806-0064",65,91350076,2944050,0,0,0,0,56884607,56884607
+1,"app-20231122005806-0064",66,107310214,6845034,0,0,0,0,54009485,30802660
+1,"app-20231122005806-0064",67,616892826,15384391,0,0,225919114,1610612544,327309150,200703525
+1,"app-20231122005806-0064",68,27971049,537798,0,0,0,0,2490659,2490659
+1,"app-20231122005806-0064",69,46042058,4577644,0,0,0,0,230540,230540
+1,"app-20231122005806-0064",70,125897308,6973368,0,0,0,0,394648654,394648654
+1,"app-20231122005806-0064",71,79952714,8152449,0,0,0,0,16612794,16612794
+1,"app-20231122005806-0064",72,450412120,6406075,0,0,0,0,13745817,13745817
+1,"app-20231122005806-0064",73,3015886980,144004979,0,0,0,0,28359,28359
+1,"app-20231122005806-0064",74,86876394,4217477,0,0,0,0,3422630,3353968
+1,"app-20231122005806-0064",75,338807438,14351282,0,0,0,0,395502728,280989779
+1,"app-20231122005806-0064",76,673641638,31676468,0,0,0,0,0,806857736
+1,"app-20231122005806-0064",77,230774052,9147656,0,0,0,0,146717217,146717217
+1,"app-20231122005806-0064",78,107846812,10276246,0,0,0,0,44007443,44007443
+1,"app-20231122005806-0064",79,144484227,26642000,0,0,0,0,110631447,110631447
+1,"app-20231122005806-0064",80,90699631,5592863,0,0,0,0,1555130,1555130
+1,"app-20231122005806-0064",81,20993626,1182922,0,0,0,0,114113,114113
+1,"app-20231122005806-0064",82,10081666,1309541,0,0,0,0,6461377,6461377
+1,"app-20231122005806-0064",83,10635646,1052388,0,0,0,0,314698,314698
+1,"app-20231122005806-0064",84,379752901,27627053,0,0,0,0,551337060,551337060
+1,"app-20231122005806-0064",85,10138894,1001791,0,0,0,0,9121,9121
+1,"app-20231122005806-0064",86,2000361,204000,0,0,0,0,5137,5137
+1,"app-20231122005806-0064",87,16826606,1859136,0,0,0,0,80021,80021
+1,"app-20231122005806-0064",88,11850097,190556,0,0,0,0,4280302,4280302
+1,"app-20231122005806-0064",89,12082957,262328,0,0,0,0,7143354,7143354
+1,"app-20231122005806-0064",90,2307042520,230615230,0,0,0,0,32093,32093
+1,"app-20231122005806-0064",91,255447123,33512991,0,0,0,0,141455590,141455590
+1,"app-20231122005806-0064",92,773059667,62783061,0,0,0,0,577545432,533658047
+1,"app-20231122005806-0064",93,777952263,63283061,0,0,0,0,577536363,546734681
+1,"app-20231122005806-0064",94,771888558,60795036,0,0,0,0,84682467,84663295
+1,"app-20231122005806-0064",95,753684875,59793317,0,0,0,0,60147359,60137773
+1,"app-20231122005806-0064",96,83308144,3598933,0,0,0,0,80158718,51145065
+1,"app-20231122005806-0064",97,185360101,11063624,0,0,0,0,393381959,393381959
+1,"app-20231122005806-0064",98,25715663,1820434,0,0,0,0,2010784,1190265
+1,"app-20231122005806-0064",99,37191527,4321976,0,0,0,0,34968781,18049658
+1,"app-20231122005806-0064",100,124770284,7634105,0,0,0,0,826,826
+1,"app-20231122005806-0064",101,59481009,1293586,0,0,0,0,57649801,57649801
+1,"app-20231122005806-0064",102,253618028,14351282,0,0,0,0,60148893,60148893
+1,"app-20231122005806-0064",103,107846812,10276246,0,0,0,0,44007442,44007442
+1,"app-20231122005806-0064",104,26406637,1176582,0,0,0,0,268419,268419
+1,"app-20231122005806-0064",105,13914431,1613923,0,0,0,0,34225786,17306663
+1,"app-20231122005806-0064",106,168292953,5553452,0,0,0,0,15644905,15644905
+1,"app-20231122005806-0064",107,30002588,1990332,0,0,0,0,7047354,7047354
+1,"app-20231122005806-0064",108,10636854,1055793,0,0,0,0,84680,84680
+1,"app-20231122005806-0064",109,17956108,958095,0,0,0,0,399113,399113
+1,"app-20231122005806-0064",110,290649244,21638115,0,0,0,0,694151,241285
+1,"app-20231122005806-0064",111,133412960,4894327,0,0,0,0,1568341,1568341
+1,"app-20231122005806-0064",112,185029300,1834412,0,0,0,0,34204441,34204441
+1,"app-20231122005806-0064",113,960918709,139595760,0,0,0,0,1568865152,1568865152
+1,"app-20231122005806-0064",114,91451667,5606846,0,0,0,0,567531,567531
+1,"app-20231122005806-0064",115,45759052,2907685,0,0,0,0,6015301,6015301
+1,"app-20231122005806-0064",116,166088899,7566199,0,0,0,0,5929781,5929781
+1,"app-20231122005806-0064",117,168836611,7668863,0,0,0,0,1108,1108
+1,"app-20231122005806-0064",118,676838724,32528525,0,0,0,0,687501157,687501157
+1,"app-20231122005806-0064",119,677897500,32630525,0,0,0,0,1375061840,687559349
+1,"app-20231122005806-0064",120,590565282,20170736,0,0,0,0,796018032,605163117
+1,"app-20231122005806-0064",121,238197152,14421315,0,0,0,0,3853246,3853246
+1,"app-20231122005806-0064",122,33316542,1831919,0,0,0,0,567132,567132
+1,"app-20231122005806-0064",123,6301099,356444,0,0,0,0,837373,837373
+1,"app-20231122005806-0064",124,33198298,2737748,0,0,0,0,1298,1298
+1,"app-20231122005806-0064",125,235001644,7784690,0,0,0,0,99876773,99876773
+1,"app-20231122005806-0064",126,893305964,45193786,0,0,0,0,1576596,1576596
