+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_cbCD7a/rapids_4_spark_profile/app-20231122005806-0064/spark_properties.csv
@@ -1,0 +1,31 @@
+propertyName,appIndex_1
+"spark.app.id","app-20231122005806-0064"
+"spark.app.initial.jar.urls","spark://10.150.30.2:45197/jars/nds-benchmark-listener-1.0-SNAPSHOT.jar"
+"spark.app.name","NDS - Power Run"
+"spark.app.startTime","1700614686297"
+"spark.app.submitTime","1700614685749"
+"spark.driver.extraJavaOptions","-Djava.net.preferIPv6Addresses=false -XX:+IgnoreUnrecognizedVMOptions --add-opens=java.base/java.lang=ALL-UNNAMED --add-opens=java.base/java.lang.invoke=ALL-UNNAMED --add-opens=java.base/java.lang.reflect=ALL-UNNAMED --add-opens=java.base/java.io=ALL-UNNAMED --add-opens=java.base/java.net=ALL-UNNAMED --add-opens=java.base/java.nio=ALL-UNNAMED --add-opens=java.base/java.util=ALL-UNNAMED --add-opens=java.base/java.util.concurrent=ALL-UNNAMED --add-opens=java.base/java.util.concurrent.atomic=ALL-UNNAMED --add-opens=java.base/sun.nio.ch=ALL-UNNAMED --add-opens=java.base/sun.nio.cs=ALL-UNNAMED --add-opens=java.base/sun.security.action=ALL-UNNAMED --add-opens=java.base/sun.util.calendar=ALL-UNNAMED --add-opens=java.security.jgss/sun.security.krb5=ALL-UNNAMED -Djdk.reflect.useDirectMethodHandle=false"
+"spark.driver.host","10.150.30.2"
+"spark.driver.memory","10G"
+"spark.driver.port","45197"
+"spark.dynamicAllocation.enabled","false"
+"spark.eventLog.dir","file:///tmp/spark-events"
+"spark.eventLog.enabled","true"
+"spark.executor.cores","12"
+"spark.executor.extraJavaOptions","-Djava.net.preferIPv6Addresses=false -XX:+IgnoreUnrecognizedVMOptions --add-opens=java.base/java.lang=ALL-UNNAMED --add-opens=java.base/java.lang.invoke=ALL-UNNAMED --add-opens=java.base/java.lang.reflect=ALL-UNNAMED --add-opens=java.base/java.io=ALL-UNNAMED --add-opens=java.base/java.net=ALL-UNNAMED --add-opens=java.base/java.nio=ALL-UNNAMED --add-opens=java.base/java.util=ALL-UNNAMED --add-opens=java.base/java.util.concurrent=ALL-UNNAMED --add-opens=java.base/java.util.concurrent.atomic=ALL-UNNAMED --add-opens=java.base/sun.nio.ch=ALL-UNNAMED --add-opens=java.base/sun.nio.cs=ALL-UNNAMED --add-opens=java.base/sun.security.action=ALL-UNNAMED --add-opens=java.base/sun.util.calendar=ALL-UNNAMED --add-opens=java.security.jgss/sun.security.krb5=ALL-UNNAMED -Djdk.reflect.useDirectMethodHandle=false"
+"spark.executor.id","driver"
+"spark.executor.instances","8"
+"spark.executor.memory","16G"
+"spark.history.fs.logDirectory","file:///tmp/spark-events"
+"spark.jars","file:///home/leey/devpub/NVIDIA/spark-rapids-benchmarks/nds/jvm_listener/target/nds-benchmark-listener-1.0-SNAPSHOT.jar"
+"spark.master","spark://dgx2h0194.spark.sjc4.nvmetal.net:7077"
+"spark.rdd.compress","True"
+"spark.repl.local.jars","file:///home/leey/devpub/NVIDIA/spark-rapids-benchmarks/nds/jvm_listener/target/nds-benchmark-listener-1.0-SNAPSHOT.jar"
+"spark.scheduler.minRegisteredResourcesRatio","1.0"
+"spark.scheduler.mode","FIFO"
+"spark.serializer.objectStreamReset","100"
+"spark.sql.adaptive.enabled","true"
+"spark.sql.broadcastTimeout","1200"
+"spark.sql.shuffle.partitions","200"
+"spark.submit.deployMode","client"
+"spark.submit.pyFiles",""
