+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_cbCD7a/rapids_4_spark_profile/app-20231122005806-0064/spark_rapids_build_info.json
@@ -1,0 +1,6 @@
+[ {
+  "sparkRapidsBuildInfo" : { },
+  "sparkRapidsJniBuildInfo" : { },
+  "cudfBuildInfo" : { },
+  "sparkRapidsPrivateBuildInfo" : { }
+} ]
