+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_cbCD7a/rapids_4_spark_profile/app-20231122005806-0064/system_properties.csv
@@ -1,0 +1,7 @@
+propertyName,appIndex_1
+"file.encoding","UTF-8"
+"java.version","17.0.8.1"
+"os.arch","amd64"
+"os.name","Linux"
+"os.version","5.15.0-1029-nvidia"
+"user.timezone","Etc/UTC"
