+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_cbCD7a/rapids_4_spark_profile/profiling_status.csv
@@ -1,0 +1,2 @@
+Event Log,Status,AppID,Description
+file:/home/leey/devpub/spark-rapids-tools/user_tools/tests/spark_rapids_tools_ut/resources/eventlogs/onprem/nds/power/eventlogs/cpu/app-20231122005806-0064.zstd,SUCCESS,app-20231122005806-0064,Took 15701ms to process
