+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_f4adD8/rapids_4_spark_profile/app-20231114200842-0001/io_metrics.csv
@@ -1,0 +1,104 @@
+appIndex,appID,sqlID,input_bytesRead_sum,input_recordsRead_sum,output_bytesWritten_sum,output_recordsWritten_sum,diskBytesSpilled_sum,memoryBytesSpilled_sum,sr_totalBytesRead_sum,sw_bytesWritten_sum
+1,"app-20231114200842-0001",24,246348669,28894693,0,0,0,0,917,917
+1,"app-20231114200842-0001",25,151502288,7567110,0,0,0,0,2384406,2384406
+1,"app-20231114200842-0001",26,758767588,24913232,0,0,0,0,249386860,169357178
+1,"app-20231114200842-0001",27,1208429,102000,0,0,0,0,0,0
+1,"app-20231114200842-0001",28,23073589,4328108,0,0,0,0,86005603,84986499
+1,"app-20231114200842-0001",29,23073589,4328108,0,0,0,0,84606480,83587314
+1,"app-20231114200842-0001",30,387077745,6314389,0,0,0,0,112252388,112252388
+1,"app-20231114200842-0001",31,25429728,991159,0,0,0,0,3540615,3540615
+1,"app-20231114200842-0001",32,19209741,1804942,0,0,0,0,616483,616483
+1,"app-20231114200842-0001",33,64446678,2267326,0,0,0,0,34260558,34260558
+1,"app-20231114200842-0001",34,558152508,14863693,0,0,0,0,422188242,422188242
+1,"app-20231114200842-0001",35,47187906,1609480,0,0,0,0,217199,217199
+1,"app-20231114200842-0001",36,66053477,1684157,0,0,0,0,13588111,13588111
+1,"app-20231114200842-0001",37,17920708,2770604,0,0,0,0,28923,16476
+1,"app-20231114200842-0001",38,26257065,4765059,0,0,0,0,6559040,6559040
+1,"app-20231114200842-0001",39,60211689,5592294,0,0,0,0,41068,41068
+1,"app-20231114200842-0001",40,151892277,7600577,0,0,0,0,4999216,4999216
+1,"app-20231114200842-0001",41,484715946,15437440,0,0,0,0,170024496,170024496
+1,"app-20231114200842-0001",42,15863884,1235454,0,0,0,0,5150,5150
+1,"app-20231114200842-0001",43,5265288,429817,0,0,0,0,2199588,2199588
+1,"app-20231114200842-0001",44,850623799,32691598,0,0,0,0,742273148,716150445
+1,"app-20231114200842-0001",45,107333089,5681975,0,0,0,0,245503,245503
+1,"app-20231114200842-0001",46,15432572,1074842,0,0,0,0,968344,968344
+1,"app-20231114200842-0001",47,83592337,5606624,0,0,0,0,3481665,3481665
+1,"app-20231114200842-0001",48,156844329,7270687,0,0,0,0,586779,586779
+1,"app-20231114200842-0001",49,744921886,30565344,0,0,0,0,331812646,331812646
+1,"app-20231114200842-0001",50,40639385,4677377,0,0,0,0,9524333,9524333
+1,"app-20231114200842-0001",51,83416205,5687683,0,0,0,0,316314,316314
+1,"app-20231114200842-0001",52,33725360,4113220,0,0,0,0,8576404,8576404
+1,"app-20231114200842-0001",53,23777008,1623319,0,0,0,0,2323105,2323105
+1,"app-20231114200842-0001",54,298257834,27724165,0,0,0,0,3957670,2494749
+1,"app-20231114200842-0001",55,155894084,18656310,0,0,0,0,42761891,42761891
+1,"app-20231114200842-0001",56,11517168,445431,0,0,0,0,12861718,11355380
+1,"app-20231114200842-0001",57,144649662,6316995,0,0,0,0,4011422,4011422
+1,"app-20231114200842-0001",58,99621890,11098429,0,0,0,0,19925,19925
+1,"app-20231114200842-0001",59,100807338,5688164,0,0,0,0,583616180,583616180
+1,"app-20231114200842-0001",60,2677682424,172805946,0,0,0,0,7515180,7515180
+1,"app-20231114200842-0001",61,54711202,1645399,0,0,0,0,28603192,28603192
+1,"app-20231114200842-0001",62,88281904,8367472,0,0,0,0,66797879,66797879
+1,"app-20231114200842-0001",63,179402315,4473469,0,0,0,0,101319,101319
+1,"app-20231114200842-0001",64,288962101,14575332,0,0,0,0,1824,1824
+1,"app-20231114200842-0001",65,78051960,3050148,0,0,0,0,50088025,50088025
+1,"app-20231114200842-0001",66,97743201,6898083,0,0,0,0,42438533,23142757
+1,"app-20231114200842-0001",67,484715946,15437440,0,0,0,0,241022656,150490249
+1,"app-20231114200842-0001",68,23600752,590847,0,0,0,0,2114484,2114484
+1,"app-20231114200842-0001",69,41910272,4577644,0,0,0,0,313801,313801
+1,"app-20231114200842-0001",70,108131488,7026417,0,0,0,0,301644773,301644773
+1,"app-20231114200842-0001",71,71797277,8205498,0,0,0,0,19095984,19095984
+1,"app-20231114200842-0001",72,340014187,6459124,0,0,0,0,11783971,11783971
+1,"app-20231114200842-0001",73,2223244336,144005000,0,0,0,0,7773,7773
+1,"app-20231114200842-0001",74,74406748,4376624,0,0,0,0,3869618,3797358
+1,"app-20231114200842-0001",75,303179396,14457380,0,0,0,0,321924366,220484772
+1,"app-20231114200842-0001",76,628563623,31676468,0,0,0,0,0,596331387
+1,"app-20231114200842-0001",77,196515365,9306803,0,0,0,0,128633229,128633229
+1,"app-20231114200842-0001",78,79442664,10329295,0,0,0,0,29433406,29433406
+1,"app-20231114200842-0001",79,143722855,26695049,0,0,0,0,100906058,100906058
+1,"app-20231114200842-0001",80,82423469,5645912,0,0,0,0,1568225,1568225
+1,"app-20231114200842-0001",81,18401764,1235971,0,0,0,0,168770,168770
+1,"app-20231114200842-0001",82,9043916,1362590,0,0,0,0,6558826,6558826
+1,"app-20231114200842-0001",83,9711616,1105437,0,0,0,0,403288,403288
+1,"app-20231114200842-0001",84,336917939,27680102,0,0,0,0,1942006575,1942006575
+1,"app-20231114200842-0001",85,9244536,1054840,0,0,0,0,16377,16377
+1,"app-20231114200842-0001",86,1552493,204000,0,0,0,0,6971,6971
+1,"app-20231114200842-0001",87,14763163,1912185,0,0,0,0,96795,96795
+1,"app-20231114200842-0001",88,11313996,243605,0,0,0,0,3990933,3990933
+1,"app-20231114200842-0001",89,11490820,315377,0,0,0,0,6507640,6507640
+1,"app-20231114200842-0001",90,1970428712,231106430,0,0,0,0,7372,7372
+1,"app-20231114200842-0001",91,213972765,33566040,0,0,0,0,85157531,85157531
+1,"app-20231114200842-0001",92,684955000,62942208,0,0,0,0,476730979,457015779
+1,"app-20231114200842-0001",93,688723320,63442208,0,0,0,0,536439909,467142805
+1,"app-20231114200842-0001",94,650903556,60901134,0,0,0,0,107874219,107866663
+1,"app-20231114200842-0001",95,634643452,59952464,0,0,0,0,86558600,86554822
+1,"app-20231114200842-0001",96,70411253,3651982,0,0,0,0,56264482,33369044
+1,"app-20231114200842-0001",97,169180869,11116673,0,0,0,0,317905595,317905595
+1,"app-20231114200842-0001",98,22708669,1919883,0,0,0,0,2363105,1342809
+1,"app-20231114200842-0001",99,31331624,4375025,0,0,0,0,31112700,16068699
+1,"app-20231114200842-0001",100,106446077,7687161,0,0,0,0,924,924
+1,"app-20231114200842-0001",101,47386982,1346635,0,0,0,0,49294096,49294096
+1,"app-20231114200842-0001",102,217984702,14457380,0,0,0,0,56368526,56368526
+1,"app-20231114200842-0001",103,79442664,10329295,0,0,0,0,29430772,29430772
+1,"app-20231114200842-0001",104,23754934,1229631,0,0,0,0,394402,394402
+1,"app-20231114200842-0001",105,12158820,1666972,0,0,0,0,30492247,15452637
+1,"app-20231114200842-0001",106,140144983,5553452,0,0,0,0,11270535,11270535
+1,"app-20231114200842-0001",107,25925461,2149479,0,0,0,0,5636197,5636197
+1,"app-20231114200842-0001",108,9711616,1108842,0,0,0,0,123041,123041
+1,"app-20231114200842-0001",109,15489076,1011144,0,0,0,0,556543,556543
+1,"app-20231114200842-0001",110,252046552,21744213,0,0,0,0,845069,296229
+1,"app-20231114200842-0001",111,122766060,4947376,0,0,0,0,1513249,1513249
+1,"app-20231114200842-0001",112,140416798,1887461,0,0,0,0,23159832,23159832
+1,"app-20231114200842-0001",113,902932010,139648809,0,0,0,0,1709609770,1709609770
+1,"app-20231114200842-0001",114,83085417,5659895,0,0,0,0,514918,514918
+1,"app-20231114200842-0001",115,42064344,2960734,0,0,0,0,5565673,5565673
+1,"app-20231114200842-0001",116,154843320,7619248,0,0,0,0,4920342,4920342
+1,"app-20231114200842-0001",117,150440529,7721912,0,0,0,0,1699,1699
+1,"app-20231114200842-0001",118,605653343,32528525,0,0,0,0,27363817,27363817
+1,"app-20231114200842-0001",119,1203223137,64306948,0,0,0,0,309814645,309812857
+1,"app-20231114200842-0001",120,533570176,20276834,0,0,0,0,651034284,481935184
+1,"app-20231114200842-0001",121,176982267,14474364,0,0,0,0,793462,793462
+1,"app-20231114200842-0001",122,31575596,1884968,0,0,0,0,646282,646282
+1,"app-20231114200842-0001",123,5127761,356444,0,0,0,0,1229769,1229769
+1,"app-20231114200842-0001",124,31158749,2790797,0,0,0,0,1737,1737
+1,"app-20231114200842-0001",125,184583708,7837739,0,0,0,0,79791287,79791287
+1,"app-20231114200842-0001",126,766755696,45194016,0,0,0,0,631950,631950
