+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_f4adD8/rapids_4_spark_profile/app-20231114200842-0001/spark_rapids_build_info.json
@@ -1,0 +1,6 @@
+[ {
+  "sparkRapidsBuildInfo" : { },
+  "sparkRapidsJniBuildInfo" : { },
+  "cudfBuildInfo" : { },
+  "sparkRapidsPrivateBuildInfo" : { }
+} ]
