+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_f4adD8/rapids_4_spark_profile/app-20231114200842-0001/spark_rapids_parameters_set_explicitly.csv
@@ -1,0 +1,6 @@
+propertyName,appIndex_1
+"spark.rapids.driver.user.timezone","Z"
+"spark.rapids.memory.host.spillStorageSize","32G"
+"spark.rapids.memory.pinnedPool.size","8g"
+"spark.rapids.sql.concurrentGpuTasks","2"
+"spark.rapids.sql.multiThreadedRead.numThreads","20"
