+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_f4adD8/rapids_4_spark_profile/app-20231114200842-0001/system_properties.csv
@@ -1,0 +1,7 @@
+propertyName,appIndex_1
+"file.encoding","UTF-8"
+"java.version","17.0.8.1"
+"os.arch","amd64"
+"os.name","Linux"
+"os.version","5.15.0-1029-nvidia"
+"user.timezone","Etc/UTC"
