+++ user_tools/tests/spark_rapids_tools_ut/resources/qualx_cache/onprem/profile/nds_local/prof_20250307013940_f4adD8/rapids_4_spark_profile/profiling_status.csv
@@ -1,0 +1,2 @@
+Event Log,Status,AppID,Description
+file:/home/leey/devpub/spark-rapids-tools/user_tools/tests/spark_rapids_tools_ut/resources/eventlogs/onprem/nds/power/eventlogs/gpu/app-20231114200842-0001.zstd,SUCCESS,app-20231114200842-0001,Took 14403ms to process
